Add expected database schema version to RPCs. 1) Client side - Minimum changes to client API. Database version   is piped around internally for use in RPC. - Schema version added to BatchOptions during   creation of batch. - RPC API change do pass a int64 schema version number   as param for all read write methods of database, table   and row. 2) Server side - schema version number is stored in databaseData object - schema version is read from db for each write operation   that needs to check version. In future leveldb is expected   to have internal caching layer making the read op   inexpensive.
MultiPart: 1/2

+++ services/syncbase/nosql/types.vdl.go
@@ -172,7 +172,7 @@ type SchemaMetadata struct {
 	// Non negative Schema version number. Should be increased with every schema change
 	// (e.g. adding fields to structs) that cannot be handled by previous
 	// versions of the app.
-	Version int64
+	Version int32
 	Policy  CrPolicy
 }
